LayoutTests/platform/wk2/imported/w3c/indexeddb/idbdatabase_deleteObjectStore4-not_reused-expected.txt


FAIL IDBDatabase.deleteObjectStore() - the object store is not reused assert_array_equals: keys property 1, expected 6 but got 1
